fix poll lint line limits
