+++ src/cli/help-command-pages.mts
@@ -1,3 +1,5 @@
+import { LOG_FILE_USAGE } from "./help-log-file-page.mts";
+
 export const COMMAND_USAGE = {
   resolve: `pr-shepherd resolve
 
@@ -182,21 +184,5 @@ Flags:
 
 Exit code: 0 on success (including no-change no-op); 1 on validation, lookup, or mutation failure.`,
 
-  "log-file": `pr-shepherd log-file
-
-Print the per-worktree append-only debug log path for the current repository.
-The log is created by the first non-help pr-shepherd command that initializes logging.
-
-Usage:
-  pr-shepherd log-file [--format text|json]
-
-Flags:
-  --format text|json   Print a raw path or {"path": "..."} JSON. Default: text.
-  --help, -h           Print this help and exit before logging setup.
-
-Environment:
-  PR_SHEPHERD_LOG_DISABLED=1 disables logging.
-  PR_SHEPHERD_STATE_DIR overrides the base state directory.
-
-Exit code: 0 on success; 1 if repository identity cannot be resolved.`,
+  "log-file": LOG_FILE_USAGE,
 } as const;
